requirements.txt
typing_extensions==4.12.2
python-dotenv==1.0.1
jsonschema==4.23.0
google-generativeai==0.8.3
couchbase==4.3.5
flask==3.1.0